Change balance table to using stored data instead of querying on mount
